requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy
psycopg2-binary==2.9.9; sys_platform != 'win32'
pydantic
pydantic-settings
redis==5.0.3
yfinance
pandas
numpy
scipy==1.14.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
httpx==0.27.0
pytest==8.3.5
python-binance==1.0.19
prometheus-fastapi-instrumentator
